+++ components/product/where/type.d.ts
@@ -1,0 +1,5 @@
+declare module 'where' {
+  export type Where<T> = {
+    [K in keyof T]: T[K]
+  }
+}
